requirements.txt
asgiref==3.5.0
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
coveralls==3.3.1
Django==3.1.14
# Django==3.2.12
django-cors-headers==3.11.0
django-currentuser==0.5.3
django-debug-toolbar==3.2.4
django-filter==21.1
django-markdownx==3.0.1
django-request-logging==0.7.2
django-simple-history==2.11.0
django-webpack-loader
djangorestframework-simplejwt>=4.4.0
djangorestframework==3.11.1
docopt==0.6.2
drf-rw-serializers==1.0.4
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.2
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Markdown==3.3.6
MarkupSafe==2.1.0
mccabe==0.6.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.7
python-coveralls==2.9.3
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0
